feat: modularize project
