+++ core/src/main/kotlin/com/rapatao/projects/ruleset/engine/context/EvalContext.kt
