+++ core/src/main/kotlin/com/rapatao/projects/ruleset/engine/types/OnFailure.kt
