+++ core/src/main/kotlin/com/rapatao/projects/ruleset/engine/types/builder/extensions/ContainsExtensions.kt
