+++ jackson/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    implementation project(":core")
+
+    implementation("com.fasterxml.jackson.core:jackson-databind:${jacksonVersion}")
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:${jacksonVersion}")
+}
