+++ kotlin-evaluator/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    api project(":core")
+}
