+++ rhino-evaluator/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    api project(":core")
+    api("org.mozilla:rhino:${rhinoVersion}")
+}
