+++ tests/src/test/kotlin/com/rapatao/projects/ruleset/engine/EvaluatorTest.kt
