+++ tests/src/test/kotlin/com/rapatao/projects/ruleset/engine/SerializationTest.kt
