+++ tests/src/test/kotlin/com/rapatao/projects/ruleset/engine/cases/OnFailureCases.kt
