+++ tests/src/test/kotlin/com/rapatao/projects/ruleset/engine/cases/OperatorWithCases.kt
