+++ tests/src/test/kotlin/com/rapatao/projects/ruleset/engine/helper/Helper.kt
