+++ tests/src/test/kotlin/com/rapatao/projects/ruleset/engine/types/ExpressionTest.kt
